refactor: move oidc discovery out of rfc module
