+++ pkgs/standards/auto_authn/auto_authn/v2/rfc8932.py
@@ -17,7 +17,7 @@ from typing import Any, Dict, List
 from fastapi import APIRouter, HTTPException, status
 
 from .runtime_cfg import settings
-from .rfc8414 import ISSUER, JWKS_PATH
+from .oidc_discovery import ISSUER, JWKS_PATH
 
 RFC8932_SPEC_URL = "https://www.rfc-editor.org/rfc/rfc8932"
